비성상적인 소리 출력 해결, 변속 출력 대응

+++ Assets/Scripts/bms/BMSHeaderParser.cs
@@ -111,8 +111,8 @@ namespace BMS
                 // STOP //
                 if (headerKey.StartsWith("#STOP"))
                 {
-                    Int32.TryParse(headerValue, out int stop);
-                    TrackInfo.stopTable.Add(Decode36(headerKey.Substring(5)), stop / 192);
+                    Double.TryParse(headerValue, out double stop);
+                    TrackInfo.stopTable.Add(Decode36(headerKey.Substring(5)), stop / 192.0);
                 }
 
                 // BPM //
