+++ Assets/Scripts/bms/models/Stop.cs
@@ -2,8 +2,8 @@ namespace BMS
 {
     public class Stop : BMSObject
     {
-        public double Key { get; private set; }
-        public Stop(int bar, double beat, double beatLength, double stop) : base(bar, beat, beatLength)
+        public int Key { get; private set; }
+        public Stop(int bar, double beat, double beatLength, int stop) : base(bar, beat, beatLength)
         {
             Key = stop;
         }
